StoragehubManager

StoragehubManager

Format

R6Class object.

Value

Object of R6Class for modelling a D4Science StoragehubManager

Note

Main user class to be used with d4storagehub4R

Methods

new(token, logger, keyring_backend)

This method is used to instantiate the StoragehubManager. By default, the url is inherited through D4Science Icproxy service.

The token is mandatory in order to use Storage Hub API.

The keyring_backend can be set to use a different backend for storing the D4science gcube token with keyring (Default value is 'env').

The logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs)

getToken()

Get user token.

getUserProfile()

Get user profile.

getUserWorkspace()

Get user workspace root path.

getWSRootID()

Get workspace root ID

getWSItemID(parentFolderID, folderPath)

Get workspace item ID

getWSElementID(parentFolderID, folderPath)

Get workspace element ID. Deprecated: use getWSItemID

listWSItems(parentFolderID)

Lists workspace items given a parentFolder ID

listWSElements(parentFolderID)

Lists workspace elements given a parentFolder ID. Deprecated: use listWSItems

listWSItemsByPath(folderPath)

Lists workspace items given a folder path

listWSElementsByPath(folderPath)

Lists workspace elements given a folder path. Deprecated: use listWSItemsByPath

searchWSItemID(itemPath)

Searchs an element ID based on an element path

searchWSFolderID(folderPath)

Searchs a folder ID based on folder path. Deprecated: use searchWSItemID

createFolder(folderPath, name, description, hidden, recursive)

Creates a folder, given a folder path, a folder name/description. By default recursive = TRUE meaning that a folder path matching nested folders will trigger all nested folders. Setting recursive = FALSE, the folder creation will work only if the folder path matches an existing folder. The hidden (default FALSE) argument can be used to set hidden folders on the workspace.

uploadFile(folderPath, file, description, archive)

Uploads a file to a folder (given a folder path). The argument description can be used to further describe the file to upload. The argument archive (default = FALSE) indicates the type of item (FILE or ARCHIVE) to be uploaded.

deleteItem(itemPath, force)

Deletes an item (folder or file) give its path on the workspace. By default force = FALSE meaning the item is moved to trash. To delete permanently the item, set force = TRUE.

getPublicFileLink(path)

Get a public link for a workspace resource

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

if (FALSE) {
  manager <- StoragehubManager$new(
    token = "<your token>",
    logger = "DEBUG"
  )
}